// nice/sdk/lakenicehandlerjs/handlers/requesthandler.d.ts
export interface IRequestHandler extends IHandlerBase {
    /**
     * handle request.
     * @param {Object} args Event arguments.
     */
    handleRequest(args: any): void;
    /**
     * get the ClientToken interface.
     * @return {IHClientToken}    The ClientToken.
    */
    getClientToken(): IHClientToken;
    /**
     * get the ImplicitToken interface.
     * @return {IHImplicitToken}    The ImplicitToken.
    */
    getImplicitToken(): IHImplicitToken;
    /**
     * get the PasswordToken interface.
     * @return {IHPasswordToken}    The PasswordToken.
    */
    getPasswordToken(): IHPasswordToken;
    /**
     * get the RefreshToken interface.
     * @return {IHRefreshToken}    The RefreshToken.
    */
    getRefreshToken(): IHRefreshToken;
    /**
     * get the UserToken interface.
     * @return {IHUserToken}    The UserToken.
    */
    getUserToken(): IHUserToken;
}
/**
 * Request handler implementation.
 */
export declare class RequestHandler implements IRequestHandler {
    lakeNiceClient: ILakeNiceClient;
    lakeNiceAuthClient: ILakeNiceAuthClient;
    requestHandlerOptions: any;
    clienttoken: IHClientToken;
    implicittoken: IHImplicitToken;
    passwordtoken: IHPasswordToken;
    refreshtoken: IHRefreshToken;
    usertoken: IHUserToken;
    addressbook: IHAddressBook;
    agents: IHAgents;
    contacts: IHContacts;
    general: IHGeneral;
    groups: IHGroups;
    lists: IHLists;
    scheduledcallbacksadmin: IHScheduledCallbacksAdmin;
    skills: IHSkills;
    agentphone: IHAgentPhone;
    chatrequestsagent: IHChatRequestsAgent;
    emails: IHEmails;
    personalcon: IHPersonalCon;
    phonecalls: IHPhoneCalls;
    scheduledcallbacksagent: IHScheduledCallbacksAgent;
    sessions: IHSessions;
    supervisor: IHSupervisor;
    voicemails: IHVoicemails;
    workitems: IHWorkItems;
    authenticate: IHAuthenticate;
    sample: IHSample;
    callback: IHCallback;
    chatrequestspatron: IHChatRequestsPatron;
    workitem: IHWorkItem;
    realtime: IHRealTime;
    reporting: IHReporting;
    wfmdata: IHWFMData;
    config: any;
    logger: any;
    parent: any;
    /**
     * Request handler api implementation.
     *
     * @param {ILakeNiceClient}		lakeNiceClient			Lake Nice client interface.
     * @param {ILakeNiceAuthClient}	lakeNiceAuthClient		Lake Nice Auth client interface.
     * @param {Object}				requestHandlerOptions	A collection of options.
     *
     * @example
     *  options = {
     *		debug: true
     *
     *  }
    */
    constructor(lakeNiceClient: ILakeNiceClient, lakeNiceAuthClient: ILakeNiceAuthClient, requestHandlerOptions: any);
    /**
     * On event.
     * @param {string} arg0 Event name.
     * @param {Object} arg1 Event function.
    */
    on(arg0: string, arg1: (event: any, val1: any, val2: any, val3: any) => void): any;
    /**
     * get the ClientToken interface.
     * @return {IHClientToken}    The ClientToken.
    */
    getClientToken(): IHClientToken;
    /**
     * get the ImplicitToken interface.
     * @return {IHImplicitToken}    The ImplicitToken.
    */
    getImplicitToken(): IHImplicitToken;
    /**
     * get the PasswordToken interface.
     * @return {IHPasswordToken}    The PasswordToken.
    */
    getPasswordToken(): IHPasswordToken;
    /**
     * get the RefreshToken interface.
     * @return {IHRefreshToken}    The RefreshToken.
    */
    getRefreshToken(): IHRefreshToken;
    /**
     * get the UserToken interface.
     * @return {IHUserToken}    The UserToken.
    */
    getUserToken(): IHUserToken;
    /**
     * handle general.
     * @param {Object} args Event arguments.
    */
    handleRequest(args: any): void;
    /**
     * Assign the request options for the specified unique id.
     *
     * @param {string} uniqueID The unique id.
     * @param {Object} requestData The request data.
     *
     * @return {Object}	The request options, either or the 'params' and 'data'.
     *
     * @example
     *  returned data = {
     *		// 'params' are the URL parameters to be sent with the request
     *		// Must be a plain object or a URLSearchParams object
     *		params: { ID: 12345 },
     *
     *		// 'data' is the data to be sent as the request body.
     *		data: { ID: 'Unique' },
     *	}
     */
    assignRequestOptions(uniqueID: string, requestData?: any): any;
}